requirements.txt
vosk
numpy
torch
pyaudio
pyserial
gigachat
ultralytics
sounddevice
opencv-contrib-python
